backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "disutils>=1.4.32.post2",
    "dotenv>=0.9.9",
    "fastapi-utils>=0.8.0",
    "fastapi[all]>=0.120.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pymongo>=4.15.3",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.0.1",
    "uvicorn>=0.38.0",
]
